# data/etc/swagger/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Pre-Ingest Tool REST API met Swagger",
		"version": "1.0.0",
		"title": "Swagger Pre-Ingest"
	},
	"host": "localhost:8000",
	"basePath": "/api/preingest",
	"schemes": [
		"http",
		"https"
	],
	"paths": {
...
		"/exporting/{guid}": {
			"post": {
				"tags": [
					"Droid metagegevens exporteren naar een CSV bestand. Vereist eerst de actie '/droid/profiling/{guid}'."
				],
				"summary": "Droid metagegevens exporteren naar een CSV bestand. Vereist eerst de actie '/droid/profiling/{guid}'.",
				"description": "Droid metagegevens exporteren naar een CSV bestand. Vereist eerst de actie '/droid/profiling/{guid}'.",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "guid",
					"in": "path",
					"description": "Map ID",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Succesful Operation. OK."
					}
				}
			}
		},
...
	}
}